PYIC-7555 TS express config
