+++ package.json
@@ -42,6 +42,7 @@
     "@types/express": "^5.0.0",
     "@types/express-session": "^1.18.0",
     "@types/mocha": "^10.0.8",
+    "@types/nunjucks": "^3.2.6",
     "@types/proxyquire": "^1.3.31",
     "@types/sinon": "^17.0.3",
     "@types/sinon-chai": "^3.2.12",
