+++ src/config/language-toggle.d.ts
@@ -1,0 +1,4 @@
+declare module "@govuk-one-login/frontend-language-toggle/build/cjs/language-param-setter.cjs" {
+  const addLanguageParam: (language: string, url: string) => void;
+  export default addLanguageParam;
+}
